+++ src/main/java/com/nasnav/dao/Product360ShopsRepository.java
@@ -26,7 +26,7 @@ public interface Product360ShopsRepository extends JpaRepository<Shop360Products
     @Query("select new com.nasnav.dto.request.ProductPositionDTO(ps.productEntity.id, ps.floor.id, ps.section.id, ps.scene.id, ps.pitch, ps.yaw, ps.productEntity.productType) " +
             " from Shop360ProductsEntity ps where ps.shopEntity.id = :id and ps.published = :published")
     List<ProductPositionDTO> findProductsPositionsFullData(@Param("id") Long id,
-                                                           @Param("published") Integer published);
+                                                           @Param("published") short published);
 
     @Query("select ps from Shop360ProductsEntity ps where ps.shopEntity.id = :id and ps.published = 1")
     List<Shop360ProductsEntity> findProductsPositionsByShopId(@Param("id") Long id);
